+++ sqlparser/lib/src/analysis/types/resolving_visitor.dart
@@ -573,7 +573,16 @@ class TypeResolver extends RecursiveVisitor<TypeExpectation, void> {
       case 'upper':
         nullableIfChildIs();
         return _textType.withoutNullabilityInfo;
+      case 'concat':
+        return _textType;
+      case 'concat_ws':
+        // null if the first argument is null
+        if (params.isNotEmpty) {
+          session._addRelation(NullableIfSomeOtherIs(e, [params.first]));
+        }
+        return _textType.withoutNullabilityInfo;
       case 'group_concat':
+      case 'string_agg':
         return _textType.withNullable(true);
       case 'date':
       case 'time':